Trait sc_block_builder::BlockBuilderApi

pub trait BlockBuilderApi<Block>: Core<Block> where
    Block: Block, {
    pub fn apply_extrinsic(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        extrinsic: <Block as Block>::Extrinsic
    ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, Self::Error> { ... }

    pub fn apply_extrinsic_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        extrinsic: <Block as Block>::Extrinsic
    ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, Self::Error> { ... }

    pub fn finalize_block(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<<Block as Block>::Header, Self::Error> { ... }

    pub fn finalize_block_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<<Block as Block>::Header, Self::Error> { ... }

    pub fn inherent_extrinsics(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        inherent: InherentData
    ) -> Result<Vec<<Block as Block>::Extrinsic, Global>, Self::Error> { ... }

    pub fn inherent_extrinsics_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        inherent: InherentData
    ) -> Result<Vec<<Block as Block>::Extrinsic, Global>, Self::Error> { ... }

    pub fn check_inherents(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        block: Block, 
        data: InherentData
    ) -> Result<CheckInherentsResult, Self::Error> { ... }

    pub fn check_inherents_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        block: Block, 
        data: InherentData
    ) -> Result<CheckInherentsResult, Self::Error> { ... }

    pub fn random_seed(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<<Block as Block>::Hash, Self::Error> { ... }

    pub fn random_seed_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<<Block as Block>::Hash, Self::Error> { ... }
}

The BlockBuilder api trait that provides the required functionality for building a block.

Provided methods

pub fn apply_extrinsic(
&self,
__runtime_api_at_param__: &BlockId<Block>,
extrinsic: <Block as Block>::Extrinsic
) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, Self::Error>

Apply the given extrinsic.

Returns an inclusion outcome which specifies if this extrinsic is included in this block or not.

pub fn apply_extrinsic_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
extrinsic: <Block as Block>::Extrinsic
) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, Self::Error>

Apply the given extrinsic.

Returns an inclusion outcome which specifies if this extrinsic is included in this block or not.

pub fn finalize_block(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<<Block as Block>::Header, Self::Error>

Finish the current block.

pub fn finalize_block_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<<Block as Block>::Header, Self::Error>

Finish the current block.

pub fn inherent_extrinsics(
&self,
__runtime_api_at_param__: &BlockId<Block>,
inherent: InherentData
) -> Result<Vec<<Block as Block>::Extrinsic, Global>, Self::Error>

Generate inherent extrinsics. The inherent data will vary from chain to chain.

pub fn inherent_extrinsics_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
inherent: InherentData
) -> Result<Vec<<Block as Block>::Extrinsic, Global>, Self::Error>

Generate inherent extrinsics. The inherent data will vary from chain to chain.

pub fn check_inherents(
&self,
__runtime_api_at_param__: &BlockId<Block>,
block: Block,
data: InherentData
) -> Result<CheckInherentsResult, Self::Error>

Check that the inherents are valid. The inherent data will vary from chain to chain.

pub fn check_inherents_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
block: Block,
data: InherentData
) -> Result<CheckInherentsResult, Self::Error>

Check that the inherents are valid. The inherent data will vary from chain to chain.

pub fn random_seed(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<<Block as Block>::Hash, Self::Error>

Generate a random seed.

pub fn random_seed_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<<Block as Block>::Hash, Self::Error>

Generate a random seed.

Loading content...