Crate sc_block_builder

Substrate block builder

This crate provides the BlockBuilder utility and the corresponding runtime api BlockBuilder.

The block builder utility is used in the node as an abstraction over the runtime api to initialize a block, to push extrinsics and to finalize a block.

Structs

BlockBuilder

Utility for building new (valid) blocks from a stream of extrinsics.

BlockBuilderBuilder

A builder for creating an instance of BlockBuilder.

BlockBuilderBuilderStage1

The second stage of the BlockBuilderBuilder.

BlockBuilderBuilderStage2

The second stage of the BlockBuilderBuilder.

BuiltBlock

A block that was build by BlockBuilder plus some additional data.

Traits

BlockBuilderApi

The BlockBuilder api trait that provides the required functionality for building a block.