Crate sc_block_builder

Substrate block builder

This crate provides the BlockBuilder utility and the corresponding runtime api BlockBuilder.

The block builder utility is used in the node as an abstraction over the runtime api to initialize a block, to push extrinsics and to finalize a block.

Structs

BlockBuilder

Utility for building new (valid) blocks from a stream of extrinsics.

BuiltBlock

A block that was build by BlockBuilder plus some additional data.

Enums

RecordProof

Used as parameter to BlockBuilderProvider to express if proof recording should be enabled.

Traits

BlockBuilderApi

The BlockBuilder api trait that provides the required functionality for building a block.

BlockBuilderProvider

Block builder provider