Trait sc_block_builder::BlockBuilderApi

pub trait BlockBuilderApi<Block>: Core<Block>where
    Block: Block,{
    fn apply_extrinsic(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        extrinsic: <Block as Block>::Extrinsic
    ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError> { ... }
    fn apply_extrinsic_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        extrinsic: <Block as Block>::Extrinsic
    ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError> { ... }
    fn apply_extrinsic_before_version_6(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        extrinsic: <Block as Block>::Extrinsic
    ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError> { ... }
    fn apply_extrinsic_before_version_6_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        extrinsic: <Block as Block>::Extrinsic
    ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError> { ... }
    fn finalize_block(
        &self,
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<<Block as Block>::Header, ApiError> { ... }
    fn finalize_block_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext
    ) -> Result<<Block as Block>::Header, ApiError> { ... }
    fn inherent_extrinsics(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        inherent: InherentData
    ) -> Result<Vec<<Block as Block>::Extrinsic, Global>, ApiError> { ... }
    fn inherent_extrinsics_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        inherent: InherentData
    ) -> Result<Vec<<Block as Block>::Extrinsic, Global>, ApiError> { ... }
    fn check_inherents(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        block: Block,
        data: InherentData
    ) -> Result<CheckInherentsResult, ApiError> { ... }
    fn check_inherents_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        block: Block,
        data: InherentData
    ) -> Result<CheckInherentsResult, ApiError> { ... }
}

The BlockBuilder api trait that provides the required functionality for building a block.

Provided Methods

fn apply_extrinsic(
&self,
__runtime_api_at_param__: &BlockId<Block>,
extrinsic: <Block as Block>::Extrinsic
) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError>

Apply the given extrinsic.

Returns an inclusion outcome which specifies if this extrinsic is included in this block or not.

fn apply_extrinsic_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
extrinsic: <Block as Block>::Extrinsic
) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError>

Apply the given extrinsic.

Returns an inclusion outcome which specifies if this extrinsic is included in this block or not.

fn apply_extrinsic_before_version_6(
&self,
__runtime_api_at_param__: &BlockId<Block>,
extrinsic: <Block as Block>::Extrinsic
) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError>

👎Deprecated

fn apply_extrinsic_before_version_6_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
extrinsic: <Block as Block>::Extrinsic
) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError>

👎Deprecated

fn finalize_block(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<<Block as Block>::Header, ApiError>

Finish the current block.

fn finalize_block_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<<Block as Block>::Header, ApiError>

Finish the current block.

fn inherent_extrinsics(
&self,
__runtime_api_at_param__: &BlockId<Block>,
inherent: InherentData
) -> Result<Vec<<Block as Block>::Extrinsic, Global>, ApiError>

Generate inherent extrinsics. The inherent data will vary from chain to chain.

fn inherent_extrinsics_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
inherent: InherentData
) -> Result<Vec<<Block as Block>::Extrinsic, Global>, ApiError>

Generate inherent extrinsics. The inherent data will vary from chain to chain.

fn check_inherents(
&self,
__runtime_api_at_param__: &BlockId<Block>,
block: Block,
data: InherentData
) -> Result<CheckInherentsResult, ApiError>

Check that the inherents are valid. The inherent data will vary from chain to chain.

fn check_inherents_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
block: Block,
data: InherentData
) -> Result<CheckInherentsResult, ApiError>

Check that the inherents are valid. The inherent data will vary from chain to chain.