Trait sc_block_builder::BlockBuilderApi

pub trait BlockBuilderApi<Block>: Core<Block>
where
    Block: Block,
{
    // Provided methods
    fn apply_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
        extrinsic: <Block as Block>::Extrinsic,
    ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError> { ... }
    fn apply_extrinsic_before_version_6(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
        extrinsic: <Block as Block>::Extrinsic,
    ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError> { ... }
    fn finalize_block(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
    ) -> Result<<Block as Block>::Header, ApiError> { ... }
    fn inherent_extrinsics(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
        inherent: InherentData,
    ) -> Result<Vec<<Block as Block>::Extrinsic>, ApiError> { ... }
    fn check_inherents(
        &self,
        __runtime_api_at_param__: <Block as Block>::Hash,
        block: Block,
        data: InherentData,
    ) -> Result<CheckInherentsResult, ApiError> { ... }
}

The BlockBuilder api trait that provides the required functionality for building a block.

Provided Methods

fn apply_extrinsic( &self, __runtime_api_at_param__: <Block as Block>::Hash, extrinsic: <Block as Block>::Extrinsic, ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError>

Apply the given extrinsic.

Returns an inclusion outcome which specifies if this extrinsic is included in this block or not.

fn apply_extrinsic_before_version_6( &self, __runtime_api_at_param__: <Block as Block>::Hash, extrinsic: <Block as Block>::Extrinsic, ) -> Result<Result<Result<(), DispatchError>, TransactionValidityError>, ApiError>

๐Deprecated

fn finalize_block( &self, __runtime_api_at_param__: <Block as Block>::Hash, ) -> Result<<Block as Block>::Header, ApiError>

Finish the current block.

fn inherent_extrinsics( &self, __runtime_api_at_param__: <Block as Block>::Hash, inherent: InherentData, ) -> Result<Vec<<Block as Block>::Extrinsic>, ApiError>

Generate inherent extrinsics. The inherent data will vary from chain to chain.

fn check_inherents( &self, __runtime_api_at_param__: <Block as Block>::Hash, block: Block, data: InherentData, ) -> Result<CheckInherentsResult, ApiError>

Check that the inherents are valid. The inherent data will vary from chain to chain.