Trait sc_block_builder::BlockBuilderProvider

pub trait BlockBuilderProvider<B, Block, RA> where
    Block: BlockT,
    B: Backend<Block>,
    Self: Sized,
    RA: ProvideRuntimeApi<Block>, {
    fn new_block_at<R: Into<RecordProof>>(
        &self, 
        parent: &BlockId<Block>, 
        inherent_digests: DigestFor<Block>, 
        record_proof: R
    ) -> Result<BlockBuilder<'_, Block, RA, B>>;

    fn new_block(
        &self, 
        inherent_digests: DigestFor<Block>
    ) -> Result<BlockBuilder<'_, Block, RA, B>>;
}

Block builder provider

Required methods

fn new_block_at<R: Into<RecordProof>>(
&self,
parent: &BlockId<Block>,
inherent_digests: DigestFor<Block>,
record_proof: R
) -> Result<BlockBuilder<'_, Block, RA, B>>

Create a new block, built on top of parent.

When proof recording is enabled, all accessed trie nodes are saved. These recorded trie nodes can be used by a third party to proof the output of this block builder without having access to the full storage.

fn new_block(
&self,
inherent_digests: DigestFor<Block>
) -> Result<BlockBuilder<'_, Block, RA, B>>

Create a new block, built on the head of the chain.